+++ autotransform/repo/base.py
@@ -10,10 +10,12 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import Any, Mapping, TypedDict
+from typing import Any, Generic, Mapping, TypedDict, TypeVar
 
 from autotransform.batcher.base import Batch
 from autotransform.repo.type import RepoType
+
+TParams = TypeVar("TParams", bound=Mapping[str, Any])
 
 
 class RepoBundle(TypedDict):
@@ -23,21 +25,21 @@ class RepoBundle(TypedDict):
     type: RepoType
 
 
-class Repo(ABC):
+class Repo(Generic[TParams], ABC):
     """The base for Repo components.
 
     Attributes:
-        params (Mapping[str, Any]): The paramaters that control operation of the Repo.
+        params (TParams): The paramaters that control operation of the Repo.
             Should be defined using a TypedDict in subclasses
     """
 
-    params: Mapping[str, Any]
+    params: TParams
 
-    def __init__(self, params: Mapping[str, Any]):
+    def __init__(self, params: TParams):
         """A simple constructor.
 
         Args:
-            params (Mapping[str, Any]): The paramaters used to set up the Repo
+            params (TParams): The paramaters used to set up the Repo
         """
         self.params = params