+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - RunnableWorker renamed to ProcessWorker
 - Runner renamed to Coordinator
 - Updated BatchMetadata to only require title
+- Add a generic to components for params so that type safety is improved
 
 ## Release 0.1.1 - Initial Beta!